Kill FMU spam
FMUs based on QTronics FMU SDK tend to log fmi2SetReal etc as INFO.
These would get printed regardless of log level.
The solution is to set our default loglevel to jm_log_level_warning and give our import context to the FMU so it is able to pick this loglevel up.