Introduce simpler MPI syntax
If MPI non-master nodes are given only a single FMU filename then it will be taken.
This removes the need for giving the entire list of FMU filenames to every node,
which would tend to grow the command line lengths quadratically.
This also removes the need to compute the number of nodes for the -np option to mpiexec:

  mpiexec -np 1 fmigo-mpi [master arguments] : -np 1 fmigo-mpi fmu1.fmu : -np 1 fmigo-mpi fmu2.fmu [: ...]